+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.137
+
+### Patch Changes
+
+- @sap-ux/fe-fpm-writer@0.43.23
+
 ## 0.1.136
 
 ### Patch Changes