+++ packages/eslint-plugin-fiori-tools/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap-ux/eslint-plugin-fiori-tools
 
+## 9.13.0
+
+### Minor Changes
+
+- 52f6549: Add .cds annotations support to enable linting of CAP apps with the eslint-plugin-fiori-tools.
+
+### Patch Changes
+
+- Updated dependencies [52f6549]
+    - @sap-ux/fiori-annotation-api@0.9.44
+
 ## 9.12.3
 
 ### Patch Changes